+++ lib/data/beans.g.dart
@@ -1,10 +1,40 @@
 // GENERATED CODE - DO NOT MODIFY BY HAND
 
-part of 'models.dart';
+part of 'beans.dart';
 
 // **************************************************************************
 // JsonSerializableGenerator
 // **************************************************************************
+
+User _$UserFromJson(Map<String, dynamic> json) {
+  return User(
+      token: json['token'] as String,
+      avatar: json['avatar'] as String,
+      firstname: json['firstname'] as String,
+      id: json['id'] as int,
+      lastname: json['lastname'] as String);
+}
+
+Map<String, dynamic> _$UserToJson(User instance) => <String, dynamic>{
+      'id': instance.id,
+      'firstname': instance.firstname,
+      'lastname': instance.lastname,
+      'avatar': instance.avatar,
+      'token': instance.token
+    };
+
+Settings _$SettingsFromJson(Map<String, dynamic> json) {
+  return Settings(
+      rememberMe: json['rememberMe'] as bool,
+      stayLoggedIn: json['stayLoggedIn'] as bool,
+      useBio: json['useBio'] as bool);
+}
+
+Map<String, dynamic> _$SettingsToJson(Settings instance) => <String, dynamic>{
+      'rememberMe': instance.rememberMe,
+      'stayLoggedIn': instance.stayLoggedIn,
+      'useBio': instance.useBio
+    };
 
 Splash _$SplashFromJson(Map<String, dynamic> json) {
   return Splash(