+++ lib/models/index.dart
@@ -1,3 +1,0 @@
-// export './splash_model.dart';
-// export './splash_ad_model.dart';
-// export './splash_guide_model.dart';