+++ lib/services/index.dart
@@ -1,3 +1,3 @@
 export './splash_service.dart';
-
+export './settings_service.dart';